Decode Liquid tokens after purify instead of via a custom URI definition
